Make sure that edit boxes have their own space visually
This patch should fix that problem that I reported on the nvda at nvda.groups.io mailing list as follows:

I have another problem with the SentenceNav add-on settings panel. This seems to be a problem that only appears visually, not in speech.

What happens is that some of the settings (from "Sentence breakers" down to "Disaple SentenceNav in applications") are all shown on top of the first setting ("Volume of chime when crossing paragraph border"). Because they are on top of each other, only the last one of them is visible.

When I check the program, it seems that all the setting elements that are created via sHelper (gui.guiHelper.BoxSizerHelper) and all the settings with sliders (which use settingSizer, the second argument to the makeSettings) get some space, but all the other settings don't get space.

Because this is a problem of on-screen layout, it may be very difficult to debug. I'm ready to help as much as I can, but any advice regarding layout of settings dialogs is appreciated.

[end quote from mailing list]

This was mostly found by looking at code of the focusHighlight add-on and some trial and error. I have tested that the controls and their labels show visually, and that one can edit and save the settings.